Funcionando

+++ TrabalhoCOM220.git/src/Reserva.java
@@ -9,13 +9,13 @@ public class Reserva implements Serializable{
     private int numeroReserva;
     private Date entrada;
     private Date saida;
-    private double desconto;
+    private int desconto;
     private ArrayList<Quarto> Quartos = new ArrayList<>();
     private Pagamento diarias;
     private Pagamento pgtReserva;
     private String Cpf;
 
-    public Reserva(int numeroReserva, Date entrada, Date saida, double desconto, Pagamento diarias, Pagamento pgtReserva, String Cpf,ArrayList<Quarto> vectorQuartos) {
+    public Reserva(int numeroReserva, Date entrada, Date saida, int desconto, Pagamento diarias, Pagamento pgtReserva, String Cpf,ArrayList<Quarto> vectorQuartos) {
         this.numeroReserva = numeroReserva;
         this.entrada = entrada;
         this.saida = saida;
@@ -50,11 +50,11 @@ public class Reserva implements Serializable{
         this.saida = saida;
     }
 
-    public double getDesconto() {
+    public int getDesconto() {
         return desconto;
     }
 
-    public void setDesconto(double desconto) {
+    public void setDesconto(int desconto) {
         this.desconto = desconto;
     }
